refactor(ui): extract bash tool renderer to tools-shell module
Slice 11b.1 step 12 of the message-part.tsx split. Pull the bash card
renderer — owns shell-output normalization + copy-to-clipboard + animated
subtitle — into its own file. Bash is shell output (not file diff), so
it deliberately doesn't share helpers with the edit / write / apply_patch
renderers waiting in the next slice.

New: packages/ui/src/components/message-part-tools-shell.tsx
  Side-effect registers: bash

Imports the shared `ShellSubmessage` primitive from
`./message-part-tool-display` rather than redeclaring it.

tsgo --noEmit GREEN. bun test: 478 pass / 3 pre-existing fail.
